+++ tests/hybrid/Hybrid.test.cpp
@@ -45,7 +45,7 @@ TEST_F(Hybrid, Case1a) {
     std::vector<T> position = { 1.75e-3, 0.75e-3 };
     std::vector<T> size = { 5e-4, 5e-4 };
     std::string stlFile = "../examples/STL/cross.stl";
-    std::unordered_map<int, arch::Opening<T>> Openings;
+    std::unordered_map<size_t, arch::Opening<T>> Openings;
     Openings.try_emplace(5, arch::Opening<T>(network->getNode(5), std::vector<T>({1.0, 0.0}), 1e-4));
     Openings.try_emplace(7, arch::Opening<T>(network->getNode(7), std::vector<T>({0.0, -1.0}), 1e-4));
     Openings.try_emplace(8, arch::Opening<T>(network->getNode(8), std::vector<T>({0.0, 1.0}), 1e-4));
